% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\name{egsea.data}
\alias{egsea.data}
\title{EGSEAdata databases information}
\usage{
egsea.data(species = "human", simple = FALSE, returnInfo = FALSE)
}
\arguments{
\item{species}{character, a species name and used to retreive the number of gene sets
for that particular species. Default is "human". Accepted values are  
"human", "homo sapiens", "hs", "mouse", "mus musculus", "mm",
"rat", "rattus norvegicus" or "rn".}

\item{simple}{logical, whether to display the number of gene sets in each 
collection or not.}

\item{returnInfo, }{logical, whether to print out the databases information or 
return it as a list.}
}
\value{
nothing.
}
\description{
It displays information about the available gene set collections in 
EGSEAdata for a species.
}
\details{
It prints out for each database: the database name, version, update/download
date, data source, supported species, gene set collections, the names of the related
 R data objects and the number of gene sets in each collection.
}
\examples{
# Example of egsea.data 
egsea.data()

}

