% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{package}
\name{EGSEAdata-package}
\alias{EGSEAdata}
\alias{EGSEAdata-package}
\title{Gene Set Collections for the EGSEA package}
\description{
This package includes gene set collections that are used for 
the Ensemble of Gene Set Enrichment Analyses (EGSEA) method for gene set testing. 
It includes Human and Mouse versions of the MSidDB (Subramanian, et al. (2005) 
PNAS, 102(43):15545-15550) and GeneSetDB (Araki, et al. (2012) FEBS Open Bio, 2:76-82) 
collections.
}
\author{
Monther Alhamdoosh, Yifang Hu and Gordon K. Smyth
}

