% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{gsetdb.all}
\alias{gsetdb.all}
\title{GeneSetDB Human Collections}
\format{list}
\source{
Araki Hiromitsu,Knapp Christoph,Tsai Peter and Print Cristin(2012), 
GeneSetDB: A comprehensive meta-database, statistical and visualisation framework 
for gene set analysis, FEBS Open Bio, 2, doi: 10.1016/j.fob.2012.04.003
}
\description{
Human gene set collections from the GeneSetDB
}

