% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{il13.data.cnt}
\alias{il13.data.cnt}
\title{Human IL-13 dataset - Raw Counts}
\format{A List object with five components: counts, group, design, contra and genes.}
\source{
The FASTQ files of this experiment are vailable from the GEO database
\url{www.ncbi.nlm.nih.gov/geo/} as series
}
\description{
It contains the raw count matrix of RNA-seq performed on samples of human normal
PBMCs, IL-13 stimulated PBMCs and IL-13R antagnonist PBBMCs. It also contains the 
contrast and design matrices of this experiment. The gene symbols mapping is also included.
}

