% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{il13.data}
\alias{il13.data}
\title{Human IL-13 dataset}
\format{A List object with two components: voom and contra.}
\source{
The count matrix of this experiment is vailable from the GEO database
\url{www.ncbi.nlm.nih.gov/geo/} as series
}
\description{
The voom object calculated from the TMM normalized count matrix of RNA-seq 
performed on samples of human normal PBMCs, IL-13 stimulated PBMCs and IL-13R 
antagnonist PBBMCs. It also contains the contrast matrix of this experiment.
}

