% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{kegg.pathways}
\alias{kegg.pathways}
\title{KEGG Pathways Collections}
\format{list}
\source{
Luo, W., Friedman, M., Shedden K., Hankenson, K. and Woolf, P GAGE: Generally Applicable
Gene Set Enrichment for Pathways Analysis. BMC Bioinformatics 2009, 10:161
}
\description{
Human, Mouse and Rat gene set collections from the KEGG database
}

