% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{mam.data}
\alias{mam.data}
\title{Mouse mammary cell dataset}
\format{A List object with two components: voom and contra.}
\source{
The count matrix of this experiment is vailable from the GEO database
\url{www.ncbi.nlm.nih.gov/geo/} as series GSE63310.
}
\description{
The voom object calculated from TMM normalized count matrix of RNA-seq
performed on samples of the epithelial cells of the mouse mammary glands 
from three populations: basal, luminal progenitor and mature luminal. It also
contains the contrast matrix of this experiment.
}

