% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{Mm.H}
\alias{Mm.H}
\title{Mouse H MSigDB Gene Set Collections}
\format{list}
\source{
http://bioinf.wehi.edu.au/software/MSigDB/
}
\description{
Mouse orthologs gene set collections from the MSigDB database Version 5
}

