% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{package}
\name{EGSEAdata-package}
\alias{EGSEAdata}
\alias{EGSEAdata-package}
\title{Gene Set Collections for the EGSEA package}
\description{
This package includes gene set collections that are used for 
the Ensemble of Gene Set Enrichment Analyses (EGSEA) method for gene set testing. 
It includes Human and Mouse versions of the MSidDB (Subramanian, et al. (2005) 
PNAS, 102(43):15545-15550) and GeneSetDB (Araki, et al. (2012) FEBS Open Bio, 2:76-82) 
collections.
}
\details{
While the gene set collections in MSigDB and GeneSetDB have different names and 
purposes, some of these collections overlap. For example, both databases contain a Gene Ontology
collection but MSigDB's collection aimed for a higher level of abstraction for the GO terms.
}
\author{
Monther Alhamdoosh, Yifang Hu and Gordon K. Smyth
}

