% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{package}
\name{EGSEAdata-package}
\alias{EGSEAdata}
\alias{EGSEAdata-package}
\title{Gene Set Collections for the EGSEA package}
\description{
This package provides R data objects for the  gene set collections that 
are used by the EGSEA package.
}
\author{
Monther Alhamdoosh, Yifang Hu and Gordon K. Smyth
}

