% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egseaData.R
\docType{data}
\name{Mm.c5}
\alias{Mm.c5}
\title{Mouse C5 MSigDB Gene Set Collections}
\format{list}
\source{
http://bioinf.wehi.edu.au/software/MSigDB/
Aravind Subramanian, Pablo Tamayo, Vamsi K. Mootha, Sayan Mukherjee, Benjamin L. Ebert, Michael A. Gillette, Amanda Paulovich, Scott L. Pomeroy, Todd R. Golub, Eric S. Lander, and Jill P. Mesirov
Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles
PNAS 2005 102 (43) 15545-15550
}
\description{
Mouse orthologs gene set collections from the MSigDB database Version 4
}

